git-svn-id: file:///var/www/svn2/oecdo/oecde@690 50b15a66-aedb-45cf-97b2-84293224aabc

+++ journal/src/main/java/org/sgu/oecde/journal/dao/JournalDAO.java
@@ -2,7 +2,6 @@ package org.sgu.oecde.journal.dao;
 
 import java.util.List;
 import java.util.Set;
-import javax.annotation.Resource;
 import org.hibernate.Criteria;
 import org.hibernate.criterion.Projections;
 import org.hibernate.criterion.Property;
@@ -32,8 +31,8 @@ public class JournalDAO extends BasicDao<EventItem> implements IJournalDao {
     public int getCountOfEvents(BaseFilter filter) throws DataAccessException {
         Criteria cr = getSession().createCriteria(type).setProjection(Projections.rowCount());
         processFilter(filter, cr);
-        List<Integer> list =  cr.setCacheable(true).list();
-        return !CollectionUtils.isEmpty(list)?list.get(0):0;
+        List<Long> list =  cr.setCacheable(true).list();
+        return !CollectionUtils.isEmpty(list)?Long.valueOf(list.get(0)).intValue():0;
     }
 
     /**
